src/third_party/blink/web_tests/external/wpt/css/cssom-view/MediaQueryList-extends-EventTarget-expected.txt

This is a testharness.js-based test.
PASS onchange adds listener
FAIL onchange removes listener assert_equals: expected 1 but got 0
FAIL listeners for "change" type are called assert_equals: expected 1 but got 0
PASS listeners with different type are not called
FAIL addEventListener "once" option is respected assert_equals: expected 1 but got 0
PASS removeEventListener removes listener
PASS dispatchEvent works as expected
Harness: the test ran to completion.
